Ensure that brace is always present
Some objects (like `MobileDeviceIosInventory`) don't have any properties
beyond `$ref`. This means they never enter the `else` case which
previously added the opening {.

By moving the brace outside of `const_from_properties`, we can make sure
it's always present and won't be added twice without some logic to check
that the brace is present or not. I think this is simpler.

The second `const_def` line is added to make the diff the same as what
is currently present, but it can probably be removed. It makes it easier
when reviewing the generated diff now to ensure that nothing has broken.